Part4: Finished 4.21 step9

+++ part4/controllers/blogs.js
@@ -46,8 +46,19 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodeToken = jwt.verify(request.token, config.SECRET)
+  if (!decodeToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() === decodeToken.id) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).send({ error: "wrong id" })
+  }
+  
 })
 
 blogsRouter.put('/:id', async (request, response) => {
